refactor: decompose opencode-aidevops plugin into focused modules (t1914)
Extract 8 modules from index.mjs (1406 → 230 lines):
- mcp-registry.mjs: MCP server catalog and registration
- config-hook.mjs: agent/MCP/provider config hook
- quality-hooks.mjs: tool execution hook wiring
- quality-logging.mjs: logging, scanning, pipeline infrastructure
- compaction.mjs: context preservation across resets
- intent-tracing.mjs: LLM intent extraction and storage
- version-tracking.mjs: opencode version drift detection
- shell-env.mjs: shell environment variable injection

index.mjs is now a thin orchestrator that imports and wires modules.
External API (AidevopsPlugin export) unchanged.
No new qlty smells introduced in extracted modules.
Removed index.mjs from qlty smells list (was complexity 223).
